+++ src/app/utilities/auth/constants.ts
@@ -48,3 +48,28 @@ export const DEFAULT_USER = {
 export const USER_TABLE_NAME = process.env.USER_TABLE_NAME || 'NDAUsers';
 export const USER_TABLE_PARTITION_ID = 'Email';
 export const DEFAULT_SIGN_IN_REDIRECT_URL = '/profile';
+
+export const OPTIONS_DATE = Array.from({ length: 31 }, (_, i) => ({
+  label: (i + 1).toString(),
+  value: i + 1,
+}));
+
+export const OPTIONS_MONTH = [
+  { label: 'Jan', value: '1' },
+  { label: 'Feb', value: '2' },
+  { label: 'Mar', value: '3' },
+  { label: 'Apr', value: '4' },
+  { label: 'May', value: '5' },
+  { label: 'Jun', value: '6' },
+  { label: 'Jul', value: '7' },
+  { label: 'Aug', value: '8' },
+  { label: 'Sep', value: '9' },
+  { label: 'Oct', value: '10' },
+  { label: 'Nov', value: '11' },
+  { label: 'Dec', value: '12' },
+];
+
+export const OPTIONS_YEAR = Array.from({ length: 99 }, (_, i) => ({
+  label: (1915 + i).toString(),
+  value: 1915 + i,
+}));
